[refactor] 인증 로직을 미들웨어로 분리하고 라우터/컨트롤러 구조 정리
